Allow adding and removing search locations for resolving #r references for Script editor. Also, fixed an issue with editor was not properly disposed when using X button on window to close and added a check to throw exception while adding document to script project if script project has already any document.
We need a way to add additional searchable paths to a script editor so that when using prefabs assembly reference locations for prefab can be added. For script project restriction for one document, this is required since if there are multiple documents in same project services like highlight service, type lookup service won't work.