Configure PrefabLoader in Singleton Scope with Ninject. Prefabs should not be used in parallel.
When running tests, each time tests run on a new task . This causes a new instance of PrefabLoader to be created over and over every time we run a new test , creating new PrefabInstance for same prefabs in their local cache. The initial thought of configuring this in ThreadScope was to allow use of Prefabs in Parallel. Using Prefabs in parallel will yield unpredictable result.