Optimize memory usage by ScriptEditor and ScriptEngine
ScriptMetadataResolver loads over 200 assemblies which are not required in ResolveMissingAssembly(). Defined a whitelist which is used by ScriptMetaDataResolver ( for script engine) now to decide what assembly can be resolved in ResolveMissingAssembly(). ScriptMetaDataResolver for ScriptEditor doesn't need to resolve anything in ResolveMissingAssembly().  ResolveReference() however will continue to resolve required references. Also, added default reference list for editors and script engines in application settings which were previously in code so that it can be changed from configuration if required. Per project additional references can also be defined.