Throw exception if Argument is not configured.
ArgumentProcessor will simply log a warning message if an argument is not configured. This can go unnoticed. It makes more sense to throw an exception if an attempt was made to get value or set value on argument using ArgumentProcessor. The components should check if argument is configured for optional arguments before trying to use the arguments with ArgumentProcessor.