Application screens for grouping controls
Applications can have a largs number of controls. Loading all the controls for application at a given time will slow down application and might not be optimal as user will usually work on a selected application screen. It is possible now to define screens for an application. Controls can be associated with these screen. Control explorer exposes a drop down now to select screen. Only controls belonging to selected screen will be loaded.