Add support for control versioning
The control details can change as the application evolves. To simplify management of test cases , it would be ideal if we don't have to edit the test cases to create and use a new control as the control details change. Instead allow revisions of control to be created and allow control version to be managed at a project level. As a result of this , we can update the version of control that should be used across the test cases in a project without editing any test cases.