Null reference exception while loading a prefab for execution
As the DesignTimePrefabLoader tries to AddSearchPaths on ScriptEditorFactory, the factory will retrieve the workspacemanager and call AddSearchPaths. However, the workspacemanager for prefab is a new and no project or documents exist yet. The metaDataReferenceResolver stays uninitialized until CreateCompilationOptions() is invoked while adding a new project to workspacemanager. Add a null check on metaDataReferenceResolver during UpdateCompilationOptions()