Infer lookup type automatically based on whether control entity is nested
ControlIdentity has a LookupType which is used to configure whether this control should be looked up relative to some other control which is already looked up. This is helpful to optimize control lookup when we want to lookup multiple controls relative to a parent control that is expensive to lookup. However, it turns out we don't need this additional configuration data to determine if the control is relative to other control . This can be inferred automatically if the control entity has a parent which is control entity itself. The control entity would consider itself relative to the parent control entity in this case.