Removed JsonIgnore attribute from PrefabEntity->Components
Pixel.Automation.Core.Components project takes a dependency on Newtonsoft so mark Components property on PrefabEntity with JsonIgnore attribute. We don't want Components to be serialized hence this is necessary. However, we are setting this collection just before prefab entity is executed and clearing it after execution is completed in a finally block to guarantee this collection will be cleared. This should safeguard against serializing Components with any items in it unless someone saves the details while this is executing. Even if this gets serialized somehow , it should not matter as the Component collection will be cleared and loaded later from prefab.