Changing the name of control while editing a control using control editor should immediately update on UI
While editing the control using control editor, if we change the name of the control in property grid, the name doesn't update on the left side of the control list which lets us select the control to edit in property grid. We need to raise notification when the name changes so that the UI can update.