Correctly replace the namepsace in script while importing a script for Prefab
While importing scripts for prefab, we need to replace the namespace of  datamodel associated with entity which is used for creating prefab with the namespace of the prefab project. This is because the datamodel of the prefab is the data model defined in the prefab project once prefab is created and not the old model.