Skip validaiton of generated input and output mapping script
Input and output mapping script can refer to the variables declared in script from the scope to which they are added. These variables can be made available in the script editor but validation fails as we no way to include these script variables while doing validation .This blocks the add prefab wizard. Skip validation for the time being until we have a better solution.