Ensure that PrefabEntity -> Components is always empty before and after execution.
Unless we clear the Components collection on PrefabEntity, this will get serialized as a part of the test case. As a result, test case fails to open if there are reference to custom types from prefab in these components. Additionally, make components field readonly and remove setter from Components property in Entity to prevent setting this from outside.