Allow multiple active version of project and prefabs
Automation project version will usually follow along the release version of target application being automated. It is possible that there can be a patch on top of an existing version of the application. This situation might require to patch the subsequent version of the automation project as well. To handle this scenario, it is now possible for projects and prefabs to have multiple active versions.