Allow version management only when project is not open in editor
If a version is published while still open in editor, we need to re-configure everything so that editor starts using the newer version. However, re-configuration is not trivial and prone to lots of issues as observed so far. To simplify things, version management is only allowed when the project is not open in editor for both automation project and prefab project.