Add application drag source on control explorer and prefab explorer
When working with an automation project, we need to go back to application explorer -> application view to drag drop the application on editor. Given we will be spending most of the time on application explorer -> control explorer view or ocassionally application explorer -> prefab explorer view , we need to avoid going to top view frequently as dropping application on editor is a frequent operation. There is a small application icon on the top left of control explorer and prefab explorer view which represents the active application and can be dropped on editor to add the application. Sadly, the icon choice is not that great :s