Changing prefab version should not require application restart
All the published prefabs have same data model assembly name. As a result, if we change the version, we can't load the assembly with same name from different path. While publishing the prefab, we compile the assembly with major version included in the name. This avoids the issue with loading when changing version.