Allow controls to be deleted
Control can be deleted now if they don't have an existing reference in any of the project. Additioinal validations are now in place to prevent editing or using a control if that is marked deleted. This can happen if a client has marked control to be deleted and another client with stale data is trying to use or edit the control.