Count of Controls and Prefabs used in  test cases and Prefab are tracked now
Whenever we add a control or prefab to an automation processs or any test case or fixture, we add a reference to it in project references file along with version used. However, this information is not enough to tell what test cases or prefabs hold these references. To answer this, we are storing usage count of each control and prefab now in test case and fixture definition file.