Fix isssue with  highlight rectangle not showing up
When the control locator is configured to show highlight rectangle, there is no highlight rectangle visible for control bounding box. This is because instance of highlight rectangle should be created on UI thread. We are not forcing initilization immediately after binding happens to ensure intance is created on UI thread.