Fix issue with custom filters not working for control lookup
GetElementMatchingCriteria is invoked with a collection of UIControl. However, different control entitie initialize filter argument with different types e.g. AutomationElement, IWebElment, etc. This causes a mismatch and script fails to execute. To adresss this we can either initialize filter argument with type of UIControl for all type of ControlEntity or each ControlEntity must ensure to invoke GetElementMatchingCriteria by converting collection of UIControl to collection of api control e.g. collection of AutomationElement.
We decided on the first approach as it provides access to all the methods availble on UIControl. However, if someone needs access to the actual API controls, they need to add additional imports and reference before they can use UIControl.GetApiControl<T> which is a bit incovenient if frequently required.