Bring TestStack/UIAComWrapper project in to Pixel.Automation
UIAComWrapper added as a project in Pixel.Automation and adjust namespace to Pixel.Windows.* from System.Windows.* to avoid conflict with UIAutomationClient library  from Microsoft.WindowsDesktop.App framework. This was done in anticipation to enhancements that might be needed in UIAComWrapper. This was referenced as a pre-built assembly at the moment.