Add supported platforms to applications
Not all plugins are supported on different operating system. As a result, when a non supported application is downloaded from an operating system that doesn't have the associated plugin available, there will be an error and application fails to start. Only applications that are supported by operating system should be downloaded