Initializer functions for Initialize environment script
It was possible to specify different initialization script earlier on a template so that different templates can use custom initialization. However, there was no way to upload custom scripts. To simplify this, we will have a single initialization script now , however, we can specify which initializer function to execute per template. At design time, we can add as many initializer functions as needed.