Dragging test data source on test cases doesn't work sometimes
Depending on the area of the test data source mouse is pressed, the data on the drop info in the drag drop handler is DataSourceConfiguration instead of TestDataSource and hence drag drop won't work. Drag drop should work irrespective of where the mouse was pressed on the test data source icon.