Allow published versions of automation project and prefabs to be opened for edit
Published versions acts as a bookmark where current state is saved and new version with the same state is created. Any major changes should happen on the new version. However, there can be scenarios where minor changes might have to be done to a version that are already published. Currently, that would require minor version to be created. However, that will duplicate all the data and if the change is very minor personal preference could be to avoid it and edit the published project instead. With this change, we will allow published versions to be edited.