Specific version of control can be modified now.
We can create revisions of control. It was possible to update configuration of most recent revision or change image on most recent revision only. When trying to configure a control or changing image on a control, there will be an option to select the version of control that needs to be modified.