Generate the data model assembly again on editing a published project or prefab
We earlier allowed published projects and prefabs to be edited. However, we will need to generate the data model assembly for the prefab or project again after it is edited as it is possible that data model could be modified. Additionally, if data model assembly of a published prefab is already loaded, it should not be possible to open it for edit.