Get latest version of test cases and fixture before opening or editing them
It is possible that a test case or fixture details might be updated e.g. name, description, control used, etc by another user. While opening a test case or fixture, we download the latest version of data files but not the details of test case or fixture itself. We need to get a latest copy of them as well so that we work on the most recent copy.