Export project data to a zip file
It is possible now to export a specific version of project data files along with dependencies (applications, controls and prefabs) to a zip file. With this, it will be possible to import this zip file in the test runner and use test runner in offline mode. Further work will be required in test runner to support this.