Don't delete generated script files if add prefab wizard is cancelled
An entry is already made in to project references file if prefab version was selected and input and output script are created. Leave them as it is if user decided to cancel the wizard after version is already confirmed. Additionally, add a safeguard for null value of input and output script while validating files while selecting version.